Extract date displaying logic to a Utils file
